requirements.txt
streamlit
PyPDF2
pypdf
pdfplumber
python-dotenv
tiktoken
langchain
langchain-community
langchain-core
beautifulsoup4
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
groq
requests
torch
numpy
nltk
openai
huggingface-hub
sentence-transformers
chromadb
scikit-learn
plotly
pandas
numpy
matplotlib

duckduckgo-search
tavily-python 